auth-api/src/tests/functional/requirements.txt
-r api.txt
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.6.1
SQLAlchemy-Utils==0.38.0
python-jose[cryptography]==3.3.0